Refactor event and tag sections to separate files
